backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "gcsfs>=2026.3.0",
    "globus-sdk>=4.5.0",
    "google-cloud-batch>=0.21.0",
    "google-cloud-logging>=3.15.0",
    "google-cloud-run>=0.16.0",
    "google-cloud-storage>=3.10.1",
    "greenlet>=3.4.0",
    "h5netcdf>=1.8.1",
    "h5py>=3.16.0",
    "modal>=1.4.1",
    "netcdf4>=1.7.4",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
    "xarray>=2026.2.0",
]
